requirements.txt
pandas
numpy
matplotlib
scikit-learn
textblob
streamlit
google-generativeai
